Award shields after chopping a forest
While here I also fixed the cost of an improvement so it is more expensive on terrain with higher movement costs (otherwise forest chops took 2 turns) and added handling for industrious civs.

+++ C7Engine/C7GameData/Save/SaveUnit.cs
@@ -15,7 +15,7 @@ namespace C7GameData.Save {
 		public string action; // "fortified"
 		public TileDirection facingDirection;
 		public string experience;
-		public int WorkerProgressTowardsJob;
+		public float WorkerProgressTowardsJob;
 		public ID WorkerJob;
 
 		// True for multiple types of automation, including worker automation
